requirements.txt
fastmcp
langchain
langchain-google-genai
python-dotenv
mcp[cli]
google
google-genai
streamlit